+++ src/diag.ts
@@ -19,6 +19,7 @@ export const enum EDiagCodeDA {
     code504 = 504, // Variadic param * >1
     code505 = 505, // param parsed Error -> unknown style
     code506 = 506, // base8 base2 diag of not support number formats
+    code507 = 507, // can't set `0xNumber` as variable , exp: `0x001 := 9999`
     // ban name
     code511 = 511, // ban var/param name same function-name.
     code512 = 512, // ban global-var name same function-name.
@@ -58,6 +59,10 @@ export const DiagsDA: DeepReadonly<TDiagsDA> = {
     506: {
         msg: 'not support of this number formats',
         path: 'https://www.autohotkey.com/docs/v1/Concepts.htm#numbers',
+    },
+    507: {
+        msg: 'can\'t set `0xNumber` as variable , exp: `0x001 := 9999`',
+        path: 'https://github.com/CoffeeChaton/vscode-autohotkey-NekoHelp/blob/main/note/code507.md',
     },
     511: {
         msg: 'var/param name same func-Name',